Move common tests from topologymanager under scope
